Update articletemplate default Author
Internally we have cases where we publish content on behalf of the client. The current configuration unfortunately shows the created-by name in the container, but the editors name in the actual article.
--
This has really bugged me - Why is the default set to show the persons name who 'published' the article and not the 'person who wrote it'? I imagine that most content writers would want credit for their work; not the editor. This pull request changes the post author to the same person as "created by" in the article, effectively giving the author the credit, not the editor.